requirements.txt
# Core Face Recognition Dependencies
face-recognition==1.3.0
opencv-python==4.11.0.86
numpy==2.3.0
dlib==20.0.0
face-recognition-models==0.3.0
Pillow==11.2.1

# Firebase and Database Dependencies
firebase-admin==6.9.0
google-cloud-firestore==2.21.0
google-cloud-storage==3.2.0
google-auth==2.35.0
google-auth-httplib2==0.2.0

# Utility Dependencies
click==8.2.1
pytz==2024.2
pandas==2.2.3

# Web Dependencies (for monitoring)
flask==3.1.0

# CSV and Data Processing
# Note: csv is a built-in Python module, no need to install

# Additional Dependencies for Enhanced Functionality
python-dateutil==2.9.0
six==1.16.0
urllib3==2.2.3
requests==2.32.3
certifi==2025.7.9

# Optional: For better performance and additional features
matplotlib==3.9.4  # For plotting and visualization
seaborn==0.13.2    # For statistical visualization
openpyxl==3.1.5    # For Excel file support

# Development and Testing
pytest==8.3.4     # For testing
flake8==7.1.1      # For code linting

# Raspberry Pi Dependencies for Solenoid Door Lock System
RPi.GPIO==0.7.1    # For GPIO control on Raspberry Pi
gpiozero==1.6.2    # High-level GPIO interface for easier control

# Raspberry Pi 5 Enhanced Dependencies
pigpio==1.78       # Advanced GPIO library with better Pi 5 support
lgpio==0.2.2.0     # New GPIO library specifically for Pi 5
rpi-lgpio==0.4     # Python bindings for lgpio (Pi 5 optimized)

# Note: Built-in modules used (no installation needed):
# - os
# - time  
# - datetime
# - sys
# - logging
